c18n: Allow stack table sealing to be optionally disabled
When creating a new thread, the stack table for that thread is allocated
and passed to libthr (via a struct) to be given to the new thread.
Previously, the stack table is sealed when passed to libthr and unsealed
(and installed) when the new thread starts. This commit allows the
sealing to be optionally disabled with CHERI_LIB_C18N_NO_OTYPE.